requirements.txt
pygame-ce
Box2D
colorama